+++ qtfred/src/mission/dialogs/ShipEditor/ShipEditorDialogModel.cpp
