+++ qtfred/src/mission/dialogs/ShipEditor/ShipGoalsDialogModel.cpp
