+++ qtfred/src/mission/dialogs/ShipEditor/ShipInitialStatusDialogModel.cpp
