+++ qtfred/src/mission/dialogs/ShipEditor/ShipSpecialStatsDialogModel.cpp
