+++ qtfred/src/mission/dialogs/ShipEditor/ShipTBLViewerModel.cpp
