+++ qtfred/src/mission/dialogs/ShipEditor/ShipTextureReplacementDialogModel.cpp
