+++ qtfred/src/ui/dialogs/ShipEditor/PlayerOrdersDialog.cpp
