+++ qtfred/src/ui/dialogs/ShipEditor/ShipFlagsDialog.cpp
