+++ qtfred/src/ui/dialogs/ShipEditor/ShipGoalsDialog.cpp
