+++ qtfred/src/ui/dialogs/ShipEditor/ShipInitialStatusDialog.cpp
