+++ qtfred/src/ui/dialogs/ShipEditor/ShipSpecialStatsDialog.cpp
