+++ qtfred/src/ui/dialogs/ShipEditor/ShipTBLViewer.cpp
