+++ qtfred/src/ui/dialogs/ShipEditor/ShipTextureReplacementDialog.cpp
